CPanel Link section
We now have the option to add a link to all kinds of categories using
the 'modify link' section --> Removed some bugs

+++ Website/AtariLegend/php/admin/links/link_cat.php
@@ -26,8 +26,19 @@ include("../../includes/common.php");
 //**************************************************************************************** 
 if (isset($action) and $action=="del_cat")
 {
-	$sql = $mysqli->query("DELETE FROM website_category WHERE website_category_id = '$category_id'") or die("Failed to delete category");
-	$smarty->assign('message', 'Category deleted');
+	//check if the categorie has some websites linked to it
+	$website_count = $mysqli->query("SELECT website_id FROM website_category_cross WHERE website_category_id = '$category_id'");
+	$nr_of_links = get_rows($website_count);
+	
+	if ($nr_of_links == 0)
+	{
+		$sql = $mysqli->query("DELETE FROM website_category WHERE website_category_id = '$category_id'") or die("Failed to delete category");
+		$smarty->assign('message', 'Category deleted');
+	}
+	else
+	{
+		$smarty->assign('message', 'Category still has websites linked to it');
+	}	
 }
 
 $website = $mysqli->query("SELECT website_category_id, website_category_name FROM website_category ORDER by website_category_name");
